requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Machine Learning
xgboost==2.0.0
lightgbm==4.0.0
tensorflow==2.13.0
keras==2.13.1

# API Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==5.0.0

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# Visualization & Reporting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Azure Integration
azure-ai-ml==1.10.0
azureml-core==1.52.0
azure-storage-blob==12.17.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.17.1
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.24.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

# Documentation
mkdocs==1.5.2
mkdocs-material==9.2.8